refactor(codemode): separate tool schema internals
