+++ packages/codemode/src/index.ts
@@ -1,4 +1,4 @@
 export * as CodeMode from "./codemode.js"
-export * as Tool from "./tool-api.js"
+export * as Tool from "./tool.js"
 export * as OpenAPI from "./openapi/index.js"
 export { ToolError, toolError } from "./tool-error.js"
