--- packages/codemode/src/tool-api.ts
@@ -1,2 +1,0 @@
-export { isDefinition, make } from "./tool.js"
-export type { Definition, JsonSchema, Options, ToolSchema as SchemaType } from "./tool.js"
